requirements.txt
streamlit
langchain
langchain-ollama
supabase
pydantic
python-dotenv
pandas
plotly
